+++ lib/moxl/src/Xec/Action/Muc/CreateChannel.php
@@ -26,6 +26,7 @@ class CreateChannel extends Action
             'name' => $this->_name,
             'nick' => $this->_nick,
             'autojoin' => $this->_autojoin,
+            'pinned' => $this->_pinned,
             'notify' => $this->_notify,
         ]);
         $this->deliver();